Refactor entry row as seperate component from actions
